[benchmarking] check the path before calling shutil.rmtree
Bug: chromium:1378275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3979479
Commit-Queue: John Chen <johnchen@chromium.org>
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#1063870}

+++ tools/perf/process_perf_results.py
@@ -787,7 +787,10 @@ def main():
         args.lightweight, args.skip_perf)
     return return_code
   finally:
-    shutil.rmtree(output_results_dir)
+    # crbug/1378275. In some cases, the temp dir could be deleted. Add a
+    # check to avoid FileNotFoundError.
+    if os.path.exists(output_results_dir):
+      shutil.rmtree(output_results_dir)
 
 
 if __name__ == '__main__':
